Updates on `command` and add new `slog` based logger (#1)
* command: add context to initConfig errors

When the `initConfig()` function fails, the previous behaviour was to
just print out the error. This made it hard to understand where the
problem was coming from since most of the errors are internal to viper
and don't provide a lot of context of when the error happened.

* command: unmarshal configuration after load

Without unmarshalling the configuration, callers need call it themselves
or rely on `viper.Get*()` calls to read configuration values. By
unmarshalling it internally we remove this requirement from users.

Use the `mapstructure.TextUnmarshallerHookFunc()` hook so configuration
fields with complex structs can provide customized unmarshalling by
implementing the `encoding.TextUnmarshaler` interface.

* log: add logger based on stdlib slog

---------

Signed-off-by: Shivansh Vij <shivanshvij@loopholelabs.io>
Co-authored-by: Shivansh Vij <shivanshvij@loopholelabs.io>